+++ src/index.js
@@ -32,7 +32,7 @@ const getQLDExposureSitesData = require("./qld-exposure-sites/index.js");
 const getSAExposureSitesData = require("./sa-exposure-sites/index.js");
 const getVicExposureSitesData = require("./vic-exposure-sites/index.js");
 const getWAExposureSitesData = require("./wa-exposure-sites/index.js");
-const getNSWCasesData = require("./nsw-cases/index.js");
+const { getNSWCasesData, getNSWVaxData } = require("./nsw-covid/index.js");
 
 // Setup some constants
 REMOTE_ROOT = "/www/dat/news/interactives/covid19-data";
@@ -172,6 +172,7 @@ const main = async () => {
   const {waExposureSites} = await getWAExposureSitesData();
 
   const { nswCases } = await getNSWCasesData()
+  const { nswVax } = await getNSWVaxData();
 
   // Format Johns Hopkins data
   const formattedJohnsHopkinsCasesData = formatJohnsHopkins(
@@ -395,6 +396,10 @@ const main = async () => {
 
   if (nswCases) {
     writeTempJSON("nsw-active-cases", nswCases);
+  }
+
+  if (nswVax) {
+    writeTempJSON("nsw-vax", nswVax);
   }
   // Also upload timestamped data with --timestamp argument
   // eg. node src/index.js --timestamp
